backend/requirements.txt
azure-identity==1.13.0b3
Flask==2.2.2
langchain==0.0.78
openai#==0.26.4
azure-search-documents==11.4.0b6
azure-storage-blob==12.14.1
python-dotenv==1.0.0
PyPDF2==3.0.1
azure-cosmos==4.4.0
Authlib==1.2.1
Werkzeug==2.2.2

pinecone-client
sentence_transformers
boto3

# for huggingface chat apis
# torch
# transformers